requirements.txt
sphinx
sphinx-toolbox
sphinx-book-theme
sphinx_design
matplotlib
numpy
scipy
tk
nltk
google-generativeai
Jinja2